+++ packages/x-adapter-platform/src/endpoint-adapters/related-tags.endpoint-adapter.ts
@@ -1,7 +1,8 @@
-import { endpointAdapterFactory } from '@empathyco/x-adapter';
+import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter';
 import { RelatedTagsRequest, RelatedTagsResponse } from '@empathyco/x-types';
 import { relatedTagsRequestMapper } from '../mappers/requests/related-tags-request.mapper';
 import { relatedTagsResponseMapper } from '../mappers/responses/related-tags-response.mapper';
+import { getBeaconServiceUrl } from './utils';
 
 /**
  * This endpoint does not support pagination in the request.
@@ -12,7 +13,8 @@ export const relatedTagsEndpointAdapter = endpointAdapterFactory<
   RelatedTagsRequest,
   RelatedTagsResponse
 >({
-  endpoint: 'https://api.{extraParams.env(.)}empathy.co/relatedtags/{extraParams.instance}',
+  endpoint: from =>
+    interpolate(`${getBeaconServiceUrl(from)}/relatedtags/{extraParams.instance}`, from),
   requestMapper: relatedTagsRequestMapper,
   responseMapper: relatedTagsResponseMapper,
   defaultRequestOptions: {