+++ packages/x-components/src/composables/__tests__/use-no-element-render.spec.ts
@@ -11,8 +11,8 @@ const renderUseNoElementRender = ({
   const wrapper = mount(
     component ??
       (defineComponent({
-        render() {
-          return useNoElementRender(this.$slots);
+        setup(props, { slots }) {
+          return () => useNoElementRender(slots);
         }
       }) as ComponentOptions<Vue>),
     {