+++ packages/x-components/src/composables/use-x-bus.ts
@@ -1,53 +1,71 @@
-import Vue, { getCurrentInstance, inject, Ref } from 'vue';
-import { XBus } from '@empathyco/x-bus';
+import { Subscription } from 'rxjs';
+import Vue, { getCurrentInstance, inject, isRef, onBeforeUnmount, Ref } from 'vue';
+import { EventPayload, SubjectPayload, XBus } from '@empathyco/x-bus';
 import { XEvent, XEventPayload, XEventsTypes } from '../wiring/events.types';
 import { WireMetadata } from '../wiring/wiring.types';
 import { getRootXComponent, getXComponentXModuleName } from '../components/x-component.utils';
 import { FeatureLocation } from '../types/origin';
-import { PropsWithType } from '../utils/types';
 import { XPlugin } from '../plugins/x-plugin';
 import { bus as xBus } from '../plugins/x-bus';
 
+interface PrivateExtendedVueComponent extends Vue {
+  xComponent?: Vue;
+}
+
 /**
- * Composable which injects the current location,
- * returns the `on` and `emit` functions from the `XBus`, applying location
- * and component metadata.
+ * Composable which injects the current location, returns the `on` and `emit` functions
+ * using the bus and applying component metadata. Also unsubscribe from events when components is
+ * unmounted.
  *
  * @remarks This composable tries to use the `XPlugin` bus and catches the exception thrown
  * by the `XPlugin` if it was not instantiated and uses the default `xBus` as fallback.
  *
  * @returns An object with the `on` and `emit` functions.
- * @public
  */
-export function useXBus(): UseXBusAPI {
+export function useXBus() {
   const injectedLocation = inject<Ref<FeatureLocation> | FeatureLocation>('location', 'none');
 
-  const currentComponent: PrivateExtendedVueComponent | undefined | null =
-    getCurrentInstance()?.proxy;
-  const currentXComponent = getRootXComponent(currentComponent ?? null);
-  if (currentComponent && currentXComponent) {
-    currentComponent.xComponent = currentXComponent;
+  const component = getCurrentInstance()?.proxy;
+  const xComponent = getRootXComponent(component ?? null);
+  if (component && xComponent) {
+    (component as PrivateExtendedVueComponent).xComponent = xComponent;
   }
+
   let bus: XBus<XEventsTypes, WireMetadata>;
   try {
     bus = XPlugin.bus;
   } catch (error) {
     bus = xBus;
   }
+
+  const subscription = new Subscription();
+  onBeforeUnmount(() => {
+    subscription.unsubscribe();
+  });
+
   return {
-    on: bus.on.bind(bus),
+    on: <Event extends XEvent, Metadata extends boolean>(event: Event, withMetadata: Metadata) => {
+      type Payload = EventPayload<XEventsTypes, Event>;
+      const observable = bus.on(event, withMetadata);
+
+      return {
+        subscribe: (
+          callback: (
+            payload: Metadata extends true ? SubjectPayload<Payload, WireMetadata> : Payload
+          ) => void
+          // Cast to any because bus.on doesn't infer conditional type referencing `withMetadata`
+        ) => subscription.add(observable.subscribe(callback as any))
+      };
+    },
     emit: <Event extends XEvent>(
       event: Event,
       payload?: XEventPayload<Event>,
       metadata: Omit<WireMetadata, 'moduleName'> = {}
     ) => {
-      const location =
-        typeof injectedLocation === 'object' && 'value' in injectedLocation
-          ? injectedLocation.value
-          : injectedLocation;
+      const location = isRef(injectedLocation) ? injectedLocation.value : injectedLocation;
 
-      bus.emit(event, payload!, createWireMetadata(metadata, currentComponent, location));
-      currentXComponent?.$emit(event, payload);
+      xComponent?.$emit(event, payload); // TODO - Pending to deprecate
+      return bus.emit(event, payload!, createWireMetadata(metadata, component, location));
     }
   };
 }
@@ -82,28 +100,4 @@ function createWireMetadata(
       }
     }
   );
-}
-
-interface PrivateExtendedVueComponent extends Vue {
-  xComponent?: Vue | undefined;
-}
-
-/**.
- * UseXBus API interface
- *
- * @public
- */
-export interface UseXBusAPI {
-  /* eslint-disable jsdoc/require-description-complete-sentence */
-  /** {@inheritDoc XBus.(on:1)} */
-  on: XBus<XEventsTypes, WireMetadata>['on'];
-  /** {@inheritDoc XBus.(emit:1)} */
-  emit(event: PropsWithType<XEventsTypes, void>): void;
-  /** {@inheritDoc XBus.(emit:2)} */
-  emit<Event extends XEvent>(
-    event: Event,
-    payload: XEventPayload<Event>,
-    metadata?: Omit<WireMetadata, 'moduleName' | 'origin' | 'location'>
-  ): void;
-  /* eslint-enable jsdoc/require-description-complete-sentence */
 }